backend-python/requirements.txt
# API e Web Scraping
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
httpx>=0.24
beautifulsoup4>=4.12
lxml>=4.9.0

# Banco de Dados
supabase>=2.0.0

# Utilitários
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
